Move tx to its own file
